% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ndexr_plot.R
\name{build_ndexr_plot}
\alias{build_ndexr_plot}
\title{build_ndexr_plot}
\usage{
build_ndexr_plot(
  hoursAgo = 24,
  timeUnit = "minutes",
  geomType = "bar",
  timeZone = "UTC",
  breaksWidth = "2 hour",
  breaksBin = 24,
  dateUnit = "5 min"
)
}
\description{
build_ndexr_plot
}
