% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge_customer.R
\name{charge_customer}
\alias{charge_customer}
\title{charge_customer}
\usage{
charge_customer(
  number = 4242424242424242,
  exp_month = 6,
  exp_year = 2026,
  cvc = 773,
  name = "Freddy Drennan",
  amount = 50,
  receipt_email = "drennanfreddy@gmail.com",
  description = "NDEXR Charge",
  live = FALSE,
  currency = "usd"
)
}
\description{
charge_customer
}
