% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_search.R
\name{data_search}
\alias{data_search}
\title{data_search}
\usage{
data_search(id = "search", username, is_admin)
}
\arguments{
\item{id}{A namespace id, shared with a server - data_search_server}

\item{username}{The username of the customer}
}
\description{
data_search
}
