% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_search.R
\name{data_search_server}
\alias{data_search_server}
\title{data_search_server}
\usage{
data_search_server(
  input,
  output,
  session,
  userSettings,
  query_results,
  user_name
)
}
\description{
data_search_server
}
