% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_generator.R
\name{query_generator_server}
\alias{query_generator_server}
\title{query_generator_server}
\usage{
query_generator_server(input, output, session, userSettings, user_name)
}
\description{
query_generator_server
}
