% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_plot.R
\name{statistics_plot_server}
\alias{statistics_plot_server}
\title{statistics_plot_server}
\usage{
statistics_plot_server(input, output, session, user_name, userSettings)
}
\description{
statistics_plot_server
}
