% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{text_right}
\alias{text_right}
\title{#' response_to_robj
#' @param url The location of the api endpoint, excluding host and port
#' @import httr
#' @import jsonlite
#' @export response_to_robj
response_to_robj <- function(url) {
host <- Sys.getenv("API_HOST")
port <- Sys.getenv("API_PORT")
url <- paste0("http://", host, ":", port, "/", url)
response <- GET(url)
response <- fromJSON(content(response, "text", encoding = "UTF-8"))
response
}
text_right}
\usage{
text_right(url, name)
}
\description{
#' response_to_robj
#' @param url The location of the api endpoint, excluding host and port
#' @import httr
#' @import jsonlite
#' @export response_to_robj
response_to_robj <- function(url) {
host <- Sys.getenv("API_HOST")
port <- Sys.getenv("API_PORT")
url <- paste0("http://", host, ":", port, "/", url)
response <- GET(url)
response <- fromJSON(content(response, "text", encoding = "UTF-8"))
response
}
text_right
}
